requirements.txt
filelock==3.13.1
fsspec==2024.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
sympy==1.12
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.10.0
